% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_model_selection.R
\name{fit_candidates}
\alias{fit_candidates}
\title{Fits candidates}
\usage{
fit_candidates(training, validation, nameY, predictors, nvmax = 8,
  nbest = 5, ID_SMA = "ID_SMA", force = NULL, exps = c(0:8)/4,
  simplify = TRUE)
}
\description{
Fits candidates
}
