% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_model_selection.R
\name{fit_variables}
\alias{fit_variables}
\title{Fits variables}
\usage{
fit_variables(training, validation, variables, predictors, nvmax = 8,
  nbest = 5, ID_SMA = "ID_SMA", force = NULL, exps = c(0:8)/4,
  type = "jpeg", output = "Spatial_residuals/Output/selection")
}
\description{
Fits variables
}
