% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{remove_col}
\alias{remove_col}
\title{Removes a column by name from a dataframe.}
\usage{
remove_col(data_frame, column_vector)
}
\arguments{
\item{data_frame}{A target dataframe}

\item{column_vector}{A string vector of column names to remove.}
}
\description{
Removes a column by name from a dataframe.
}
