% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{select_leaps_model}
\alias{select_leaps_model}
\title{Retrieves the lm object from a leaps object by index.}
\usage{
select_leaps_model(leaps_object, model_index, x, y)
}
\arguments{
\item{leaps_object}{The leaps object to retrieve from.}

\item{model_index}{The model index to retrieve.}

\item{x}{The predictors used to construct the leaps_object}

\item{y}{A vector of the predicted variable}
}
\description{
Retrieves the lm object from a leaps object by index.
}
